# docs/swagger.yaml
openapi: 3.0.3
info:
  title: Naabu API - Descoberta de Exposições de Serviço
  description: |
    API para descoberta e validação de exposições de serviço em larga escala.
    
    ## Funcionalidades
    - Port scanning com Naabu
    - Detecção de vulnerabilidades com 9 probes especializados
    - CVE Detection automático com Nuclei v3 SDK
    - Deep scanning automático com Nmap NSE
    - Suporte a IPs, hostnames e notação CIDR
    
    ## Probes Disponíveis
    - **FTP**: Detecta servidores com login anônimo
    - **VNC**: Identifica VNC sem autenticação
    - **RDP**: Detecta RDP com criptografia fraca
    - **LDAP**: Identifica LDAP com bind anônimo
    - **PPTP**: Detecta VPN legacy vulnerável
    - **rsync**: Identifica módulos rsync acessíveis
    - **SSH Cipher**: Detecta SSH com cifras fracas (CBC, 3DES, arcfour)
    - **SSH MAC**: Detecta SSH com MACs fracos (MD5, SHA1-96, RIPEMD)
    - **CVE Detection**: Detecta CVEs HIGH/CRITICAL usando Nuclei v3 templates
  version: 2.0.0
  contact:
    name: API Support
    email: support@naabu-api.local
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:9082
    description: Local development server (updated port)
  - url: http://localhost:8082
    description: Local development server (legacy)
  - url: http://localhost:8080
    description: Docker container
paths:
  /scan:
    post:
      summary: Criar novo job de scan
      description: |
        Inicia um novo job de scan assíncrono. O scan será executado em background
        através do sistema de worker pools.
      operationId: createScanJob
      tags:
        - Scanning
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ScanRequest'
            examples:
              single-ip:
                summary: Scan de IP único
                value:
                  ips: ["192.168.1.1"]
                  ports: "80,443,22"
                  enable_probes: true
                  enable_deep_scan: true
              hostname:
                summary: Scan de hostname
                value:
                  ips: ["example.com"]
                  ports: "80,443"
                  enable_probes: true
                  enable_deep_scan: false
              cidr:
                summary: Scan de rede CIDR
                value:
                  ips: ["192.168.1.0/24"]
                  ports: "22,3389"
                  enable_probes: true
                  enable_deep_scan: true
              multiple:
                summary: Múltiplos alvos
                value:
                  ips: ["8.8.8.8", "google.com", "10.0.0.0/30"]
                  ports: "53,80,443"
                  enable_probes: true
                  enable_deep_scan: true
      responses:
        '202':
          description: Job criado com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AsyncScanResponse'
              example:
                scan_id: "550e8400-e29b-41d4-a716-446655440000"
                status: "queued"
                message: "Job criado com sucesso"
        '400':
          description: Requisição inválida
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                invalid-ip:
                  value:
                    error: "IP inválido: 999.999.999.999"
                    request_id: "req-123"
                empty-list:
                  value:
                    error: "Lista de IPs não pode estar vazia"
                    request_id: "req-124"
        '500':
          description: Erro interno do servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/v1/scans:
    get:
      summary: Listar todos os scans
      description: |
        Retorna uma lista paginada de todos os scans com suporte a filtros e ordenação.
        Inclui informações resumidas de cada scan como status, duração, IPs, portas e estatísticas.
      operationId: listScans
      tags:
        - Scanning
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [queued, running, completed, failed, cancelled]
          description: Filtrar por status do scan
          example: completed
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
          description: Número de itens por página
          example: 10
        - name: offset
          in: query
          schema:
            type: integer
            minimum: 0
            default: 0
          description: Posição inicial (para paginação)
          example: 0
        - name: sort_by
          in: query
          schema:
            type: string
            enum: [created_at, updated_at, status, id]
            default: created_at
          description: Campo para ordenação
          example: created_at
        - name: sort_desc
          in: query
          schema:
            type: boolean
            default: true
          description: Ordenação decrescente
          example: true
      responses:
        '200':
          description: Lista de scans
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListScansResponse'
              example:
                scans:
                  - id: 1
                    scan_id: "550e8400-e29b-41d4-a716-446655440000"
                    status: "completed"
                    ips: ["192.168.1.1", "example.com"]
                    ports: "80,443,22"
                    created_at: "2025-07-02T10:30:00Z"
                    updated_at: "2025-07-02T10:32:15Z"
                    completed_at: "2025-07-02T10:32:15Z"
                    duration_ms: 135000
                    total_ports: 6
                    open_ports: 3
                    vulnerable_ports: 1
                    probes_run: 2
                pagination:
                  current_page: 1
                  per_page: 20
                  total_items: 45
                  total_pages: 3
                  has_next: true
                  has_prev: false
        '400':
          description: Parâmetros inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/scans/{scan_id}:
    get:
      summary: Buscar scan específico por ID
      description: |
        Retorna informações detalhadas de um scan específico incluindo resultados completos,
        probe results, deep scan artifacts e estatísticas completas.
      operationId: getScanById
      tags:
        - Scanning
      parameters:
        - name: scan_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: ID único do scan
          example: "550e8400-e29b-41d4-a716-446655440000"
      responses:
        '200':
          description: Detalhes completos do scan
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobStatusResponse'
              example:
                scan_id: "550e8400-e29b-41d4-a716-446655440000"
                status: "completed"
                created_at: "2025-07-02T10:30:00Z"
                updated_at: "2025-07-02T10:32:15Z"
                completed_at: "2025-07-02T10:32:15Z"
                results:
                  results:
                    - ip: "192.168.1.1"
                      ports:
                        - port: 22
                          protocol: "tcp"
                          state: "open"
                          service_name: "ssh"
                  summary:
                    total_ips: 2
                    total_ports: 6
                    open_ports: 3
                    vulnerable_ports: 1
                    probes_run: 2
                    deep_scans_run: 0
                    duration_ms: 135000
                    errors: 0
        '400':
          description: ID do scan inválido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Scan não encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
    ScanRequest:
      type: object
      required:
        - ips
      properties:
        ips:
          type: array
          items:
            type: string
          minItems: 1
          maxItems: 100
          description: |
            Lista de alvos para scan. Suporta:
            - IPs individuais (ex: 192.168.1.1)
            - Hostnames (ex: example.com)
            - Notação CIDR (ex: 192.168.1.0/24)
          example: ["192.168.1.1", "example.com", "10.0.0.0/30"]
        ports:
          type: string
          description: |
            Portas para scan. Formatos suportados:
            - Portas individuais: "80,443,22"
            - Ranges: "1-1000"
            - Mix: "22,80,443,1000-2000"
            - Vazio/null: usa portas padrão
          example: "21,22,23,25,53,80,110,111,135,139,143,443,445,993,995,1723,3306,3389,5432,5900,8080"
        enable_probes:
          type: boolean
          default: true
          description: Habilita execução de probes de vulnerabilidade
        enable_deep_scan:
          type: boolean
          default: true
          description: Habilita deep scan com Nmap quando vulnerabilidades são encontradas
    AsyncScanResponse:
      type: object
      properties:
        scan_id:
          type: string
          format: uuid
          description: ID único do job de scan
          example: "550e8400-e29b-41d4-a716-446655440000"
        status:
          type: string
          enum: [queued, running, completed, failed, cancelled]
          description: Status atual do job
          example: "queued"
        message:
          type: string
          description: Mensagem descritiva
          example: "Job criado com sucesso"
        request_id:
          type: string
          description: ID da requisição para tracking
          example: "req-123456"
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Mensagem de erro
          example: "IP inválido: 999.999.999.999"
        request_id:
          type: string
          description: ID da requisição para tracking
          example: "req-123456"
        details:
          type: object
          description: Detalhes adicionais do erro (opcional)
...
    ListScansResponse:
      type: object
      properties:
        scans:
          type: array
          items:
            $ref: '#/components/schemas/ScanJobSummary'
          description: Lista de scans
        pagination:
          $ref: '#/components/schemas/PaginationInfo'
      required:
        - scans
        - pagination
...
    PaginationInfo:
      type: object
      properties:
        current_page:
          type: integer
          description: Página atual
        per_page:
          type: integer
          description: Itens por página
        total_items:
          type: integer
          description: Total de itens
        total_pages:
          type: integer
          description: Total de páginas
        has_next:
          type: boolean
          description: Há próxima página
        has_prev:
          type: boolean
          description: Há página anterior
      required:
        - current_page
        - per_page
        - total_items
        - total_pages
        - has_next
        - has_prev
...
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key para autenticação (ainda não implementado)
...